refactor cron to enqueue documents for background processing
